Renamed the class token not tokens
